Testing Body Frame Velocities

+++ include/lib/differential_drivetrain.hpp
@@ -10,7 +10,7 @@ class DifferentialDrivetrain : public Drivetrain {
         friend class Chassis;
         
     public:
-        DifferentialDrivetrain(pros::MotorGroup *leftMotors, pros::MotorGroup *rightMotors, double wheelDiameter, double wheelTrackWidth, double gearRatio)
+        DifferentialDrivetrain(pros::MotorGroup *leftMotors, pros::MotorGroup *rightMotors, float wheelDiameter, float wheelTrackWidth, float gearRatio)
         : Drivetrain(wheelDiameter, wheelTrackWidth, gearRatio), leftMotors(leftMotors), rightMotors(rightMotors) {}
 
         DifferentialDrivetrain() : Drivetrain(0, 0, 0), leftMotors(nullptr), rightMotors(nullptr) {}
