+++ include/util/pose.hpp
@@ -3,12 +3,12 @@
 
 class Pose {
     private:
-        double x; // in inches
-        double y; // in inches
-        double theta; // in radians 
+        float x; // in inches
+        float y; // in inches
+        float theta; // in radians 
 
     public:
-        Pose(double x, double y, double theta) : x(x), y(y), theta(theta) {}
+        Pose(float x, float y, float theta) : x(x), y(y), theta(theta) {}
         Pose() : x(0), y(0), theta(0) {}
 
         /**
@@ -16,13 +16,13 @@ class Pose {
          */
         void reset() { x = 0; y = 0; theta = 0; }
 
-        double getX() const { return x; }
-        double getY() const { return y; }
-        double getTheta() const { return theta; }
+        float getX() const { return x; }
+        float getY() const { return y; }
+        float getTheta() const { return theta; }
 
-        void setX(double x) { this->x = x; }
-        void setY(double y) { this->y = y; }
-        void setTheta(double theta) { this->theta = theta; }
+        void setX(float x) { this->x = x; }
+        void setY(float y) { this->y = y; }
+        void setTheta(float theta) { this->theta = theta; }
 
 
         /**
@@ -31,7 +31,7 @@ class Pose {
          * @param y The new y-coordinate.
          * @param theta The new orientation (in radians).
         */
-        void setPose(double x, double y, double theta) {
+        void setPose(float x, float y, float theta) {
             this->x = x;
             this->y = y;
             this->theta = theta;
@@ -48,21 +48,21 @@ class Pose {
          * @param other The other pose to calculate the distance to.
          * @return The distance in inches.
          */
-        double distanceTo(const Pose& other);
+        float distanceTo(const Pose& other);
 
         /**
          * @brief Calculates the angle to another pose.
          * @param other The other pose to calculate the angle to.
          * @return The angle in radians.
          */
-        double angleTo(const Pose& other);
+        float angleTo(const Pose& other);
 
         /**
          * @brief Rotates the pose by a given angle.
          * @param angle The angle to rotate by (in radians).
          * @return A new Pose object representing the rotated pose.
          */
-        Pose rotate(double angle);
+        Pose rotate(float angle);
 
         std::string to_string();
 };
